requirements.txt
requests~=2.24.0
python-telegram-bot~=12.8
python-json-logger~=2.0.0
loguru~=0.5.3